Move `EditableSettingControl` into `settings`
