+++ crates/settings/src/settings.rs
@@ -1,3 +1,4 @@
+mod editable_setting_control;
 mod keymap_file;
 mod settings_file;
 mod settings_store;
@@ -7,6 +8,7 @@ use rust_embed::RustEmbed;
 use std::{borrow::Cow, str};
 use util::asset_str;
 
+pub use editable_setting_control::*;
 pub use keymap_file::KeymapFile;
 pub use settings_file::*;
 pub use settings_store::{
