+++ crates/settings_ui/Cargo.toml
@@ -14,7 +14,6 @@ path = "src/settings_ui.rs"
 [dependencies]
 command_palette_hooks.workspace = true
 feature_flags.workspace = true
-fs.workspace = true
 gpui.workspace = true
 project.workspace = true
 settings.workspace = true
