+++ crates/settings_ui/src/theme_settings_ui.rs
@@ -1,38 +1,8 @@
-use fs::Fs;
 use gpui::AppContext;
 use project::project_settings::{InlineBlameSettings, ProjectSettings};
-use settings::{update_settings_file, Settings};
+use settings::{EditableSettingControl, Settings};
 use theme::ThemeSettings;
 use ui::{prelude::*, CheckboxWithLabel, NumericStepper};
-
-/// A UI control that can be used to edit a setting visually.
-pub trait EditableSettingControl: RenderOnce {
-    /// The type of the setting value.
-    type Value: Send;
-
-    /// The settings type to which this setting belongs.
-    type Settings: Settings;
-
-    /// Returns the name of this setting.
-    fn name(&self) -> SharedString;
-
-    /// Returns a new instance of this setting.
-    fn new(cx: &AppContext) -> Self;
-
-    /// Applies the given setting file to the settings file contents.
-    ///
-    /// This will be called when writing the setting value back to the settings file.
-    fn apply(settings: &mut <Self::Settings as Settings>::FileContent, value: Self::Value);
-
-    /// Writes the given setting value to the settings files.
-    fn write(value: Self::Value, cx: &AppContext) {
-        let fs = <dyn Fs>::global(cx);
-
-        update_settings_file::<Self::Settings>(fs, cx, move |settings, _cx| {
-            Self::apply(settings, value);
-        });
-    }
-}
 
 #[derive(IntoElement)]
 pub struct UiFontSizeSetting(Pixels);
